Add files via upload
In this code i created a function that compares each element in a 1-D array. The array is generated using the random function from numpy. Comparing each element allows it to determine when to buy/sell the fictional stock that is generated by the random generator.

To generate the 1-D array, I used a Rayleigh distribution which is available in numpy, for the scale of the distribution, I used the choice function also from numpy to generate it.